Reorganization of files and folders for ComputeSimulator and fetching files from compiler_output
